# web/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Ea.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/systems": {
      "get": {
        "tags": [
          "Systemer"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/LifecycleStatus"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SystemType"
            }
          },
          {
            "name": "teamId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "businessOwnerId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Systemer"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/systems/parent-candidates": {
      "get": {
        "tags": [
          "Systemer"
        ],
        "parameters": [
          {
            "name": "forSystemId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SystemRef"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/systems/{id}": {
      "get": {
        "tags": [
          "Systemer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Systemer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Systemer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/systems/{id}/confirm": {
      "post": {
        "tags": [
          "Systemer"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmSystemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmSystemRequest": {
        "required": [
          "version"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": [
              "null",
              "integer"
            ],
            "format": "uint32"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "LifecycleStatus": {
        "enum": [
          "Planlagt",
          "Indfases",
          "IDrift",
          "Udfases",
          "Nedlagt"
        ]
      },
...
      "SystemDetail": {
        "required": [
          "id",
          "name",
          "aliases",
          "description",
          "type",
          "lifecycleStatus",
          "managingTeam",
          "parent",
          "modules",
          "roles",
          "createdAt",
          "updatedAt",
          "lastConfirmedAt",
          "lastConfirmedByName",
          "version",
          "permissions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemType"
              }
            ]
          },
          "lifecycleStatus": {
            "$ref": "#/components/schemas/LifecycleStatus"
          },
          "managingTeam": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TeamDto"
              }
            ]
          },
          "parent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemRef"
              }
            ]
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleDto"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleAssignmentDto"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastConfirmedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastConfirmedByName": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          },
          "permissions": {
            "$ref": "#/components/schemas/SystemPermissions"
          }
        }
      },
...
      "SystemListItem": {
        "required": [
          "id",
          "name",
          "parent",
          "matchedAlias",
          "type",
          "lifecycleStatus",
          "managingTeam",
          "businessOwner",
          "lastConfirmedAt",
          "moduleCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "parent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemRef"
              }
            ]
          },
          "matchedAlias": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemType"
              }
            ]
          },
          "lifecycleStatus": {
            "$ref": "#/components/schemas/LifecycleStatus"
          },
          "managingTeam": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TeamDto"
              }
            ]
          },
          "businessOwner": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PersonDto"
              }
            ]
          },
          "lastConfirmedAt": {
            "type": "string",
            "format": "date-time"
          },
          "moduleCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SystemListResponse": {
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemListItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SystemRef": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "SystemType": {
        "enum": [
          "Saas",
          "Standardsystem",
          "Egenudviklet",
          "LowCode",
          "Platform",
          "LokalLoesning",
          null
        ]
      },
      "SystemWriteRequest": {
        "required": [
          "name",
          "aliases",
          "description",
          "type",
          "lifecycleStatus",
          "managingTeamId",
          "parentSystemId",
          "roles",
          "version"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "aliases": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SystemType"
              }
            ]
          },
          "lifecycleStatus": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LifecycleStatus"
              }
            ]
          },
          "managingTeamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "parentSystemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "roles": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/RoleAssignmentInput"
            }
          },
          "version": {
            "type": [
              "null",
              "integer"
            ],
            "format": "uint32"
          }
        }
      },
...
    }
  },
...
}